cpu: x64: prelu: fix address && memory sanitizers warnings memory - possbile address to uninitialized data_type var address - possible access to dangling Xbyak:Address

+++ src/cpu/x64/prelu/jit_prelu_base_kernel.hpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright 2020 Intel Corporation
+* Copyright 2020-2021 Intel Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -57,7 +57,7 @@ private:
     size_t calc_tail_size(const memory_desc_wrapper &tensor_md) const noexcept;
     const memory_desc_wrapper tensor_md_;
     const size_t number_vmm_single_compute_ = 0;
-    int number_reserved_vmms_ = 0;
+    size_t number_reserved_vmms_ = 0;
 };
 
 } // namespace x64
